+++ demo/components/victory-brush-demo.js
@@ -2,7 +2,7 @@
 import React from "react";
 import { range, merge, random } from "lodash";
 import {
-  VictoryChart, VictoryZoom, VictoryArea, VictoryLine, VictoryGroup, VictoryAxis, VictoryStack
+  VictoryChart, VictoryZoomContainer, VictoryArea, VictoryLine, VictoryGroup, VictoryAxis, VictoryStack
 } from "../../src/index";
 import { VictoryTheme } from "victory-core";
 
@@ -80,14 +80,9 @@ export default class App extends React.Component {
       <div className="demo">
         <h1>VictoryZoom</h1>
 
-        <VictoryZoom>
-          <VictoryGroup style={{parent: parentStyle}} data={this.state.transitionData}>
-            <VictoryLine style={{data: this.state.style}} />
-          </VictoryGroup>
-        </VictoryZoom>
 
-        <VictoryZoom>
           <VictoryChart style={{parent: parentStyle}}
+            containerComponent={<VictoryZoomContainer/>}
             scale={{
               x: "time"
             }}
@@ -113,23 +108,23 @@ export default class App extends React.Component {
               ]}
             />
           </VictoryChart>
-        </VictoryZoom>
 
-        <VictoryZoom>
-          <VictoryChart style={{parent: parentStyle}} animate={{duration: 1500}}>
+          <VictoryChart style={{parent: parentStyle}}
+            animate={{duration: 1500}}
+            containerComponent={<VictoryZoomContainer/>}
+          >
             <VictoryLine
               style={{parent: parentStyle, data: this.state.style}}
               data={this.state.data}
-              label={"label\none"}
             />
           </VictoryChart>
-        </VictoryZoom>
 
         <button onClick={() => this.setState({zoomDomain: this.getZoomDomain()})}>
           New domain
         </button>
-        <VictoryZoom zoomDomain={this.state.zoomDomain}>
+
           <VictoryChart
+            containerComponent={<VictoryZoomContainer zoomDomain={this.state.zoomDomain}/>}
             animate={{duration: 1500}}
             style={{parent: parentStyle}}
             events={[{
@@ -156,10 +151,8 @@ export default class App extends React.Component {
               sample={25}
             />
           </VictoryChart>
-        </VictoryZoom>
 
-        <VictoryZoom>
-          <VictoryChart style={{parent: parentStyle}}>
+          <VictoryChart style={{parent: parentStyle}} containerComponent={<VictoryZoomContainer/>}>
             <VictoryLine
               style={{
                 parent: parentStyle,
@@ -189,10 +182,8 @@ export default class App extends React.Component {
               y={(d) => d * d}
             />
           </VictoryChart>
-        </VictoryZoom>
 
-        <VictoryZoom>
-          <VictoryChart style={{parent: parentStyle}}>
+          <VictoryChart style={{parent: parentStyle}} containerComponent={<VictoryZoomContainer/>}>
             <VictoryArea
               style={{parent: parentStyle, data: {stroke: "#333", fill: "#888", opacity: 0.4}}}
               data={this.state.data}
@@ -202,10 +193,9 @@ export default class App extends React.Component {
             <VictoryLine data={this.state.data} interpolation="stepBefore"/>
             <VictoryAxis dependentAxis/>
           </VictoryChart>
-        </VictoryZoom>
 
-        <VictoryZoom>
           <VictoryChart style={{parent: parentStyle}}
+            containerComponent={<VictoryZoomContainer/>}
             theme={VictoryTheme.material}
             events={[{
               childName: "area-1",
@@ -266,7 +256,6 @@ export default class App extends React.Component {
             </VictoryStack>
             <VictoryAxis dependentAxis/>
           </VictoryChart>
-        </VictoryZoom>
       </div>
     );
   }
