+++ src/components/containers/zoom-helper-methods.js
